add validation and check authorization for the logged in user
at every endpoint, I checked about the token to ensure that the user is logged in, also in update & delete posts, the user must be the authorithed one to do these features

+++ src/routes/postRouter.ts
@@ -1,4 +1,4 @@
-import { postService } from '../modules';
+import { postService, userService } from '../modules';
 import express from 'express';
 import { multerUpload } from '../middlewares/Multer';
 import { cloudinaryInstance } from '../modules/Cloudinary/services/Cloudinary';
@@ -6,6 +6,7 @@ import { handleMulterError } from '../middlewares/Multer';
 import { Request, Response } from 'express';
 import { endpoint } from '../core/endpoint';
 import { CustomError } from '../middlewares';
+import { secretAccessKey } from '../config';
 
 const router = express.Router();
 
@@ -25,8 +26,18 @@ router.post(
       }
       photoURL = imageURL;
     }
-    const userId = Number(req.body.userId as unknown as string);
-    const { description } = req.body;
+    const token = req.cookies.accessToken;
+    if (!token) {
+      throw new CustomError('Unauthorized', 401);
+    }
+    const user = await userService.authenticateUser(token, secretAccessKey);
+    if (!user) {
+      throw new CustomError('Unauthorized', 401);
+    }
+    const userId = user.id;
+    const description = postService.createPostSchema.parse({
+      description: req.body.description,
+    }).description;
     const newpostObject = await postService.createOne({
       description,
       photoURL,
@@ -39,6 +50,10 @@ router.post(
 router.get(
   '/',
   endpoint(async (req, res) => {
+    const token = req.cookies.accessToken;
+    if (!token) {
+      throw new CustomError('Unauthorized', 401);
+    }
     const postObjects = await postService.findMany();
     if (!postObjects) {
       throw new CustomError('Posts not found', 404);
@@ -50,6 +65,10 @@ router.get(
 router.get(
   '/:id',
   endpoint(async (req, res) => {
+    const token = req.cookies.accessToken;
+    if (!token) {
+      throw new CustomError('Unauthorized', 401);
+    }
     const id = Number(req.params.id);
     const postObject = await postService.findOne({ id });
     if (!postObject) {
@@ -62,17 +81,43 @@ router.get(
 router.put(
   '/:id',
   multerUpload.single('image'),
+  handleMulterError,
   endpoint(async (req, res) => {
     const id = Number(req.params.id);
-    const { description, photoURL } = req.body;
     const posts = await postService.findOne({ id });
     if (!posts) {
       throw new CustomError('Post not found', 404);
     }
+    let photoURL;
+    if (req.file) {
+      const localFilePath = req.file.path;
+      const { isSuccess, imageURL } = await cloudinaryInstance.uploadImage(
+        localFilePath,
+      );
+      if (!isSuccess) {
+        throw new Error();
+      }
+      photoURL = imageURL;
+    }
+    const token = req.cookies.accessToken;
+    if (!token) {
+      throw new CustomError('Invalid Credentials', 401);
+    }
+    const user = await userService.authenticateUser(token, secretAccessKey);
+    if (!user) {
+      throw new CustomError('Invalid Credentials', 401);
+    }
+    const userId = user.id;
+    if (posts.userId !== userId) {
+      throw new CustomError('Unauthorized', 401);
+    }
+    const updatePostSchema = postService.updatePostSchema.parse({
+      description: req.body.description,
+    }).description;
     const updatedpostObject = await postService.updateOne(
       { id },
       {
-        description,
+        description: updatePostSchema,
         photoURL,
       },
     );
@@ -83,10 +128,22 @@ router.put(
 router.delete(
   '/:id',
   endpoint(async (req, res) => {
+    const token = req.cookies.accessToken;
+    if (!token) {
+      throw new CustomError('Unauthorized', 401);
+    }
+    const user = await userService.authenticateUser(token, secretAccessKey);
+    if (!user) {
+      throw new CustomError('Unauthorized', 401);
+    }
     const id = Number(req.params.id);
     const deletedpostObject = await postService.deleteOne({ id });
     if (!deletedpostObject) {
       throw new CustomError('Post not found', 404);
+    }
+    const userId = user.id;
+    if (deletedpostObject.userId !== userId) {
+      throw new CustomError('Unauthorized', 401);
     }
     res.status(200).json(deletedpostObject);
   }),
